+++ src/main/java/de/automatic/ui/colorconfiguraiton/entities/Sample.java
@@ -3,22 +3,21 @@ package de.automatic.ui.colorconfiguraiton.entities;
 public interface Sample {
 
 	public double get(Channels c);
-	
+
 	public double getC1();
 
-	public void setC1(int r);
+	public void setC1(double r);
 
 	public double getC2();
 
-	public void setC2(int g);
+	public void setC2(double g);
 
 	public double getC3();
 
-	public void setC3(int b);
+	public void setC3(double b);
 
 	public int getCount();
 
 	public void setCount(int count);
-	
-	
+
 }
